Add AudioFrame serialization
This CL allows AudioFrames to be serialized and posted to a worker. The
model is the same as for VideoFrame serialization, with the difference
that we copy data, instead of passing a reference around.

There is a copy when serializing the data, and another copy each time
the data is deserialized. In practice, copying audio shouldn't be too
expensive, but adding a shareable backing to blink::AudioBuffer would
be nice.

Bug: 1168418
Change-Id: I9d1a272fddb9595ea0f5fd2b28b38b92a5acc241