URLPattern: Compile regexp with `u` flag.
This CL enables unicode mode when compiling the internal regexp.  This
is the recommended mode for new features.  One benefit we get here is
that the regexp compilation will fail for unknown escape sequences which
helps allow future escape sequences to be added to the standard.

It turns out, however, that path-to-regexp was over-escaping its
regular expressions.  So we had to remove `=`, `!`, and `:` from the
list of characters to escape.

This has all been discussed on:

https://github.com/WICG/urlpattern/issues/25

Bug: 1141510
Change-Id: Id4eb6ceaa3b4aa0bb9bc33e200cd40d0b65feb49