Apply SVG mask color filter during recording
This patch applies the SVG mask type during recording rather than
storing it as a color filter on the paint property node to be applied
later. This fixes a bug where composited masks with luminance (the
default for SVG) did not work because color filters did not apply
properly when used with blend mode.

Bug: 1187487
Change-Id: I1a19accbc2ffb2b6b8d4535ea2a86950e21e96e0