[composite-bgcolor-animation] Ensure ElementID for transparent background
When we try to start a background color animation on the
compositor thread, the required UniqueId can be missing.

This CL fixes the problem by ensuring that the UniqueId
exists for the layout object that has background color
animation. In this way, it should fix all the cases where
we might miss a repaint.

Bug: 1184832
Change-Id: I0a4a675cbe979660de13867defb10b612edbbef1